Update2 FlowPbPbpikp.cxx

+++ PWGCF/Flow/Tasks/FlowPbPbpikp.cxx
@@ -9,16 +9,16 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include <TProfile.h>
-#include <TRandom3.h>
-#include "TList.h"
+#include <cmath>
 #include <iostream>
 #include <utility>
 #include <array>
-#include <cmath>
 #include <string>
 #include <vector>
 
+#include <TProfile.h>
+#include <TRandom3.h>
+#include "TList.h"
 #include <CCDB/BasicCCDBManager.h>
 #include "Framework/runDataProcessing.h"
 #include "Framework/AnalysisTask.h"
@@ -39,7 +39,6 @@
 
 #include "Common/Core/trackUtilities.h"
 #include "Common/DataModel/Multiplicity.h"
-#include "Common/DataModel/PIDResponse.h"
 #include "CommonConstants/PhysicsConstants.h"
 #include "Framework/AnalysisDataModel.h"
 #include "Framework/StepTHn.h"
@@ -65,7 +64,7 @@ struct testgf{
   O2_DEFINE_CONFIGURABLE(cfgCutEta, float, 0.8f, "Eta range for tracks")
   O2_DEFINE_CONFIGURABLE(cfgCutChi2prTPCcls, float, 2.5, "Chi2 per TPC clusters")
   O2_DEFINE_CONFIGURABLE(cfgUseNch, bool, false, "Use Nch for flow observables")
-  O2_DEFINE_CONFIGURABLE(cfgNbootstrap, int, 10, "Number of subsamples")
+  O2_DEFINE_CONFIGURABLE(cfgNbootstrap, int16_t, 10, "Number of subsamples")
 
   ConfigurableAxis axisVertex{"axisVertex", {20, -10, 10}, "vertex axis for histograms"};
   ConfigurableAxis axisPhi{"axisPhi", {60, 0.0, constants::math::TwoPI}, "phi axis for histograms"};
